feat: 상태 저장
