+++ packages/mobile/src/__mocks__/constructionList.ts
@@ -1,0 +1,9 @@
+const constructionList = [
+  { id: 1, name: "건물", path: "/place/school/1" },
+  { id: 2, name: "식당", path: "/place/school/2" },
+  { id: 3, name: "기숙사", path: "/place/school/3" },
+  { id: 4, name: "명소", path: "/place/school/4" },
+  { id: 5, name: "프린터기", path: "/place/school/5" },
+];
+
+export default constructionList;
