+++ packages/mobile/src/__mocks__/foodList.ts
@@ -1,0 +1,8 @@
+const foodList = [
+  { id: 1, name: "한식", path: "/place/food/1" },
+  { id: 2, name: "중식", path: "/place/food/2" },
+  { id: 3, name: "일식", path: "/place/food/3" },
+  { id: 4, name: "양식", path: "/place/food/4" },
+];
+
+export default foodList;
