+++ packages/mobile/src/__mocks__/imageList.ts
@@ -1,0 +1,34 @@
+const imageList = [
+  {
+    id: 1,
+    name: "충북대학교 사회과학대학",
+    description: "충청북도 청주시 흥덕구 사창동 470",
+    src: "https://search.pstatic.net/common/?autoRotate=true&quality=95&type=w750&src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20180130_263%2F1517301583613dwuaL_JPEG%2FZh5SeEjUT12rWxcLJ2nstPaB.jpg",
+  },
+  {
+    id: 2,
+    name: "충북대학교 사회과학대학",
+    description: "충청북도 청주시 흥덕구 사창동 470",
+    src: "https://search.pstatic.net/common/?autoRotate=true&quality=95&type=w750&src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20180130_263%2F1517301583613dwuaL_JPEG%2FZh5SeEjUT12rWxcLJ2nstPaB.jpg",
+  },
+  {
+    id: 3,
+    name: "충북대학교 사회과학대학",
+    description: "충청북도 청주시 흥덕구 사창동 470",
+    src: "https://search.pstatic.net/common/?autoRotate=true&quality=95&type=w750&src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20180130_263%2F1517301583613dwuaL_JPEG%2FZh5SeEjUT12rWxcLJ2nstPaB.jpg",
+  },
+  {
+    id: 4,
+    name: "충북대학교 사회과학대학",
+    description: "충청북도 청주시 흥덕구 사창동 470",
+    src: "https://search.pstatic.net/common/?autoRotate=true&quality=95&type=w750&src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20180130_263%2F1517301583613dwuaL_JPEG%2FZh5SeEjUT12rWxcLJ2nstPaB.jpg",
+  },
+  {
+    id: 5,
+    name: "충북대학교 사회과학대학",
+    description: "충청북도 청주시 흥덕구 사창동 470",
+    src: "https://search.pstatic.net/common/?autoRotate=true&quality=95&type=w750&src=https%3A%2F%2Fldb-phinf.pstatic.net%2F20180130_263%2F1517301583613dwuaL_JPEG%2FZh5SeEjUT12rWxcLJ2nstPaB.jpg",
+  },
+];
+
+export default imageList;
