+++ packages/mobile/src/__mocks__/index.ts
@@ -1,1 +1,8 @@
-export {};
+import constructionList from "./constructionList";
+import foodList from "./foodList";
+import imageList from "./imageList";
+import menuList from "./menuList";
+import playList from "./playList";
+import snackList from "./snackList";
+
+export { constructionList, foodList, imageList, menuList, playList, snackList };
