+++ packages/mobile/src/__mocks__/menuList.ts
@@ -1,0 +1,9 @@
+const menuList = [
+  { id: 1, name: "학교", path: "/place/school" },
+  { id: 2, name: "식사", path: "/place/food" },
+  { id: 3, name: "편리", path: "/place/convenient" },
+  { id: 4, name: "간식", path: "/place/snack" },
+  { id: 5, name: "놀거리", path: "/place/play" },
+];
+
+export default menuList;
