+++ packages/mobile/src/__mocks__/playList.ts
@@ -1,0 +1,8 @@
+const playList = [
+  { id: 1, name: "PC방", path: "/place/play/1" },
+  { id: 2, name: "코인노래방", path: "/place/play/2" },
+  { id: 3, name: "즉석사진", path: "/place/play/3" },
+  { id: 4, name: "기타", path: "/place/play/4" },
+];
+
+export default playList;
