+++ packages/mobile/src/__mocks__/snackList.ts
@@ -1,0 +1,6 @@
+const snackList = [
+  { id: 1, name: "카페", path: "/place/snack/1" },
+  { id: 2, name: "아이스크림", path: "/place/snack/2" },
+];
+
+export default snackList;
