+++ packages/mobile/src/components/templates/PlaceTemplate/style.module.scss
@@ -1,0 +1,142 @@
+@import "@shared/styles/color.scss";
+@import "@shared/styles/variables.scss";
+@import "@shared/styles/mixin.scss";
+
+.header {
+  position: fixed;
+  top: 0;
+  right: 0;
+  left: 0;
+  z-index: 100;
+  display: flex;
+  justify-content: space-between;
+  height: 70px;
+  background-color: $white;
+}
+
+.link {
+  padding: 26px 22px 25px 21px;
+  display: block;
+  line-height: 0;
+}
+
+.title {
+  font-size: 18px;
+  padding: 24px 0 13px;
+  letter-spacing: 0.02em;
+  font-weight: 500;
+  line-height: 23px;
+  color: $black-100;
+
+  @include ellipsis;
+}
+
+.place_link {
+  color: $black-40;
+  padding: 26px 10px;
+  font-size: 14px;
+  line-height: 18px;
+  background-color: inherit;
+  flex: 0 0 auto;
+}
+
+.menu {
+  position: fixed;
+  top: 70px;
+  left: 0;
+  right: 0;
+  z-index: 100;
+  padding: 12px 21px 12px 13px;
+  background-color: $white;
+  border-top: 3px solid $background;
+}
+
+.list {
+  display: flex;
+
+  @include flicking;
+}
+
+.menu_link {
+  display: block;
+  font-size: 14px;
+  line-height: 18px;
+  padding: 8px;
+  white-space: nowrap;
+  color: $black-80;
+
+  &:not(:first-child) {
+    margin-left: 16px;
+  }
+
+  &[aria-selected="true"] {
+    background-color: $red-50;
+    color: $white;
+    border-radius: 100px;
+
+    &:hover {
+      @include hover-status($white, none);
+    }
+  }
+
+  &[aria-selected="false"] {
+    &:hover {
+      @include hover-status(inherit, none);
+    }
+  }
+}
+
+.content {
+  padding: 11px 15px 0 13px;
+  background-color: $background;
+}
+
+.image_list {
+  margin-top: 190px;
+}
+
+.item {
+  display: flex;
+  flex-direction: row;
+  justify-content: left;
+  align-items: inherit;
+  background-color: $white;
+  padding: 15px 35px 15px 15px;
+  border-radius: 20px;
+
+  &:not(:first-child) {
+    margin-top: 11px;
+  }
+
+  &:hover {
+    text-decoration: none;
+  }
+}
+
+.school_image {
+  width: 90px;
+  height: 90px;
+  border-radius: 20px;
+  flex: 0 0 auto;
+}
+
+.summary {
+  flex: 1 1 auto;
+
+  &:not(:first-child) {
+    margin-left: 23px;
+  }
+}
+
+.summary_title {
+  display: block;
+  padding-top: 20px;
+  color: $black-100;
+  line-height: 18px;
+}
+
+.summary_description {
+  display: block;
+  margin-top: 10px;
+  color: $black-40;
+}
