+++ packages/mobile/src/page/Map/index.tsx
@@ -83,7 +83,7 @@ function Map() {
               <PlaceArrow className={$.arrow} />
             </div>
           )}
-          <NavLink to="/place?category=1" className={$.link}>
+          <NavLink to="/place/school" className={$.link}>
             <PlaceMenu className={$.icon} />
             <span className="blind">장소탐색하기</span>
           </NavLink>
